// src/shared/types/mock-server-data.ts
import { Coords } from './coords.type.js';

export type MockServerData = {
  titles: string[]
  descriptions: string[]
  dates: Date[]
  cities: string[]
  previews: string[]
  photos: string[]
  coords: Coords[]
  emails: string[]
  avatars: string[]
  passwords: string[]
  names: string[]
};